+++ src/settings/MappingProfiles/MappingProfilesTransformationsModal/TransformationsField/TransformationsField.js
@@ -1,13 +1,9 @@
-import React, {
-  forwardRef,
-  useMemo,
-} from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 import { Field } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
-import AutoSizer from 'react-virtualized-auto-sizer';
-import { FixedSizeList as List } from 'react-window';
+
 import { isEqual } from 'lodash';
 
 import {
@@ -16,8 +12,7 @@ import {
   Checkbox,
 } from '@folio/stripes/components';
 
-import { Row } from './Row';
-import { HeaderRow } from './HeaderRow';
+import { MultiColumnList } from '../../../../components/MultiColumnList';
 
 const columnWidths = {
   isSelected: '5%',
@@ -26,15 +21,6 @@ const columnWidths = {
 };
 const visibleColumns = ['isSelected', 'fieldName', 'transformation'];
 
-const outerElementType = forwardRef((props, ref) => (
-  <div
-    id="mapping-profiles-form-transformations"
-    ref={ref}
-    role="grid"
-    {...props}
-  />
-));
-
 export const TransformationField = React.memo(({
   contentData,
   isSelectAllChecked,
@@ -42,6 +28,8 @@ export const TransformationField = React.memo(({
   onSelectAll,
 }) => {
   const intl = useIntl();
+  const headerRowHeight = 40;
+  const rowHeight = 50;
 
   const formatter = useMemo(() => ({
     isSelected: record => (
@@ -101,6 +89,13 @@ export const TransformationField = React.memo(({
     transformation: intl.formatMessage({ id: 'ui-data-export.mappingProfiles.transformations.transformation' }),
   }), [intl, isSelectAllChecked, onSelectAll, selectAllLabel]);
 
+  const itemsData = useMemo(() => ({
+    contentData,
+    formatter,
+    visibleColumns,
+    columnWidths,
+  }), [contentData, formatter]);
+
   return (
     <FieldArray
       name="transformations"
@@ -108,49 +103,12 @@ export const TransformationField = React.memo(({
       subscription={fieldArraySubscription}
     >
       {() => (
-        !contentData.length
-          ? <span data-test-list-empty>{intl.formatMessage({ id: 'stripes-components.tableEmpty' })}</span>
-          : (
-            <AutoSizer>
-              {({
-                height,
-                width,
-              }) => (
-                <List
-                  outerElementType={outerElementType}
-                  itemCount={contentData.length + 1}
-                  itemSize={50}
-                  width={width}
-                  height={height}
-                >
-                  {({
-                    index,
-                    style,
-                  }) => (!index
-                    ? (
-                      <HeaderRow
-                        key="headerRow"
-                        style={style}
-                        columnMapping={columnMapping}
-                        visibleColumns={visibleColumns}
-                        columnWidths={columnWidths}
-                      />
-                    ) : (
-                      <Row
-                        key={contentData[index - 1].displayNameKey}
-                        style={style}
-                        isOdd={Boolean(index % 2)}
-                        rowData={contentData[index - 1]}
-                        formatter={formatter}
-                        visibleColumns={visibleColumns}
-                        columnWidths={columnWidths}
-                      />
-                    ))
-                  }
-                </List>
-              )}
-            </AutoSizer>
-          )
+        <MultiColumnList
+          headerRowHeight={headerRowHeight}
+          rowHeight={rowHeight}
+          itemsData={itemsData}
+          columnMapping={columnMapping}
+        />
       )}
     </FieldArray>
   );