+++ test/bigtest/helpers/OverlayContainer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export const OverlayContainer = () => (
+  <div
+    id="OverlayContainer"
+    style={{
+      width: '100%',
+      height: '100%',
+      position: 'absolute',
+      zIndex: 2,
+    }}
+  />
+);