+++ test/bigtest/helpers/index.js
@@ -1,2 +1,5 @@
-export * from './setup-application';
+export * from './setupApplication';
 export * from './prefixKeys';
+export * from './getTotalSelectedMessage';
+export * from './OverlayContainer';
+export * from './translationsProperties';