+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Provide translations for the mapping profile transformation field names. UIDEXP-151.
 * Integrate mapping profiles transformations list with backend. UIDEXP-92.
 * Accommodate UI to the change for the retrieving of job logs API endpoint. UIDEXP-145.
+* Implement mapping profile transformations editing. UIDEXP-131.
 
 ## [2.0.1](https://github.com/folio-org/ui-data-export/tree/v2.0.1) (2020-07-09)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.0...v2.0.1)